& other 64 bit fixes
2005-01-08  Not Zed  <NotZed@Ximian.com>

        * libedataserver/e-xml-hash-utils.c (e_xmlhash_write): cast field
        length to int.

        * libedataserver/e-url.c (e_url_shroud): setup string manually
        don't use *.%s as it depends on locale.

+++ libedataserver/e-xml-hash-utils.c
@@ -250,7 +250,7 @@ e_xmlhash_write (EXmlHash *hash)
 	
 	filesave = g_alloca (strlen (hash->filename) + 5);
 	if ((slash = strrchr (hash->filename, '/')))
-		sprintf (filesave, "%.*s.#%s", slash - hash->filename + 1, hash->filename, slash + 1);
+		sprintf (filesave, "%.*s.#%s", (int)(slash - hash->filename + 1), hash->filename, slash + 1);
 	else
 		sprintf (filesave, ".#%s", hash->filename);
